+++ cpfoujsTMP.js
@@ -1,0 +1,5 @@
+
+mkdir ./static/examplesFou.jsexe
+
+cp -R ./dist-ghcjs/build/x86_64-linux/ghcjs-0.2.1/examplesFou-0.1.0.1/c/examplesFou/build/examplesFou/examplesFou.jsexe/* ./static/examplesFou.jsexe/
+