+++ cpw3jsTMP.sh
@@ -1,0 +1,6 @@
+
+
+mkdir ./static/examplesW3.jsexe
+
+cp -R ./dist-ghcjs/build/x86_64-linux/ghcjs-0.2.1/examplesW3-0.1.0.1/c/examplesW3/build/examplesW3/examplesW3.jsexe/* ./static/examplesW3.jsexe/
+