+++ examplesSem/app/Main.hs
@@ -1,0 +1,9 @@
+
+module Main where
+
+import MainW (mainW)
+import Reflex.Dom (run)
+
+main :: IO ()
+main = run mainW
+