+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Revision history for reflex-dom-themes
 
+## 0.1.0.1  -- 2018-01-28
+
+* Bootstrap.css and W3.css minor version updates.
+* reflex-dom-htmlea minor version update
+* New "project"-stuff and associated default.nix and cabal-files.
+* Example-updates to show new table-component (of relfex-dom-htmlea)
+* Other small changes.
 
 ## 0.1.0.0  -- 2017-10-09